Update code w static code analysis fixes and rename triangleClassification.py to triangle_classification.py
This code includes the updated code with the fixes made based on the problems reported by the Pylint static code analyzer. One of those problems was that the name of the python file, triangleClassification.py, did not adhere to snake_case naming style, so please note that this commit also include updating that to triangle_classification.py as well.